+++ ext/posix/glob.c
@@ -27,8 +27,11 @@
 Find all files in this directory matching a shell pattern.
 @function glob
 @string[opt="*"] pat shell glob pattern
-@param flags currently limited to posix.glob.GLOB_MARK
-@treturn table matching filenames
+@param flags bitwise inclusive OR of zero or more of `GLOB_ERR`, `GLOB_MARK`
+ and `GLOB_NOCHECK`
+@treturn[1] table matching filenames, if successful
+@treturn[2] nil
+@treturn[2] one of `GLOB_ABORTED`, `GLOB_NOMATCH` or `GLOB_NOSPACE`
 @see glob(3)
 @see glob.lua
 */
@@ -37,11 +40,16 @@ Pglob(lua_State *L)
 {
 	const char *pattern = optstring(L, 1, "*");
 	int flags = checkint(L, 2);
+	int globret = 0;
 	glob_t globres;
 
 	checknargs(L, 2);
-	if (glob(pattern, flags, NULL, &globres))
-		return pusherror(L, pattern);
+	if ((globret = glob(pattern, flags, NULL, &globres)) != 0)
+	{
+		lua_pushnil(L);
+		lua_pushinteger(L, globret);
+		return 2;
+	}
 	else
 	{
 		unsigned int i;
@@ -73,6 +81,11 @@ Constants.
 Glob constants.
 @table posix.glob
 @int GLOB_MARK append slashes to matches that are directories.
+@int GLOB_ERR instead of ignoring non-readable directories, return GLOB_ABORTED
+@int GLOB_ABORTED encountered a non-readable directory
+@int GLOB_NOCHECK return the original pattern if there are no matches
+@int GLOB_NOMATCH pattern does not match any existing pathname
+@int GLOB_NOSPACE not enough memory to continue
 @usage
   -- Print glob constants supported on this host.
   for name, value in pairs (require "posix.glob") do
@@ -89,7 +102,12 @@ luaopen_posix_glob(lua_State *L)
 	luaL_newlib(L, posix_glob_fns);
 	lua_pushstring(L, LPOSIX_VERSION_STRING("glob"));
 	lua_setfield(L, -2, "version");
+	LPOSIX_CONST( GLOB_ERR );
 	LPOSIX_CONST( GLOB_MARK );
+	LPOSIX_CONST( GLOB_NOCHECK );
+	LPOSIX_CONST( GLOB_ABORTED );
+	LPOSIX_CONST( GLOB_NOMATCH );
+	LPOSIX_CONST( GLOB_NOSPACE );
 
 	return 1;
 }